separate session handling into concern
